+++ src/app/data/situation.ts
